+++ src/pages/student/suggestion/ui/BenefitAllPage.tsx
@@ -1,36 +1,34 @@
 // 내역 전체보기 페이지 — 특정 월의 모든 혜택 내역을 스크롤로 확인한다
 
 import { Ionicons } from "@expo/vector-icons";
-import { router } from "expo-router";
-import { useState } from "react";
+import { router, useLocalSearchParams } from "expo-router";
 import { Pressable, ScrollView, Text, View } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
 import { colorTokens } from "@/shared/styles/tokens";
-import type { Benefit } from "../model/types";
+import { PageLayout } from "@/shared/ui/layout/PageLayout";
 import { getMonthFromDateStr } from "../model/types";
+import { useBenefitsStore } from "../model/useBenefitsStore";
+import { useMonthNavigator } from "../model/useMonthNavigator";
 import { BenefitList } from "./BenefitList";
 import { BenefitSummary } from "./BenefitSummary";
 import { MonthNavigator } from "./MonthNavigator";
 
-interface BenefitAllPageProps {
-	initialMonth: number;
-	benefits: Benefit[];
-}
-
-export function BenefitAllPage({
-	initialMonth,
-	benefits,
-}: BenefitAllPageProps) {
-	const [month, setMonth] = useState(initialMonth);
-	const handlePrev = () => setMonth((m) => (m === 1 ? 12 : m - 1));
-	const handleNext = () => setMonth((m) => (m === 12 ? 1 : m + 1));
+export function BenefitAllPage() {
+	const { month: monthParam } = useLocalSearchParams<{ month: string }>();
+	const initialMonth = monthParam
+		? parseInt(monthParam, 10)
+		: new Date().getMonth() + 1;
+	const { month, handlePrev, handleNext } = useMonthNavigator(initialMonth);
+	const benefits = useBenefitsStore((s) => s.benefits);
 
 	const benefitsForMonth = benefits.filter(
 		(b) => getMonthFromDateStr(b.date) === month,
 	);
 
 	return (
-		<SafeAreaView edges={["top"]} className="flex-1 bg-content-inverse">
+		<PageLayout
+			className="flex-1 bg-content-inverse"
+			contentContainerClassName="flex-1"
+		>
 			{/*뒤로가기 버튼 + 제휴 내역 타이틀 */}
 			<View className="flex-row items-center px-screen-m py-3 gap-[5px]">
 				<Pressable onPress={() => router.back()} hitSlop={8}>
@@ -61,6 +59,6 @@ export function BenefitAllPage({
 					<BenefitList benefits={benefitsForMonth} />
 				</ScrollView>
 			</View>
-		</SafeAreaView>
+		</PageLayout>
 	);
 }
